requirements.txt
streamlit==1.25.0          # For building the app interface
pandas==2.0.3              # For handling data operations
numpy==1.24.4              # Ensures compatibility with Pandas
matplotlib==3.7.2          # For creating plots and visualizations
seaborn==0.12.2            # For advanced statistical visualizations
scikit-learn==1.3.1        # For machine learning operations like LabelEncoder
Pillow==10.0.1             # For handling images
beautifulsoup4==4.12.2     # For web scraping
requests==2.31.0           # For making HTTP requests
pyarrow==12.0.1            # Required by Streamlit for caching
gensim==4.3.1              # For topic modeling and natural language processing
xgboost==1.7.6             # For gradient boosting machine learning algorithms
pycaret==3.0.2             # For low-code machine learning automation
catboost==1.2.0            # For categorical boosting algorithms
optuna==3.3.0              # For hyperparameter optimization
